# Conflicts:
#	src/main/java/com/emc/ecs/servicebroker/repository/ServiceInstance.java
#	src/main/java/com/emc/ecs/servicebroker/service/BucketInstanceWorkflow.java
#	src/main/java/com/emc/ecs/servicebroker/service/EcsService.java
#	src/test/java/com/emc/ecs/servicebroker/service/BucketInstanceWorkflowTest.java

+++ src/main/java/com/emc/ecs/servicebroker/service/EcsServiceInstanceService.java
@@ -3,8 +3,10 @@ package com.emc.ecs.servicebroker.service;
 import com.emc.ecs.servicebroker.EcsManagementClientException;
 import com.emc.ecs.servicebroker.model.PlanProxy;
 import com.emc.ecs.servicebroker.model.ServiceDefinitionProxy;
+import com.emc.ecs.servicebroker.repository.LastOperationSerializer;
 import com.emc.ecs.servicebroker.repository.ServiceInstance;
 import com.emc.ecs.servicebroker.repository.ServiceInstanceRepository;
+import com.emc.object.s3.S3Exception;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,20 +18,24 @@ import org.springframework.cloud.servicebroker.service.ServiceInstanceService;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
 
+import java.io.IOException;
 import java.util.Map;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionException;
 
 import static java.lang.String.format;
 
 @Service
 public class EcsServiceInstanceService implements ServiceInstanceService {
 
-    private static final Logger logger = LoggerFactory.getLogger(EcsServiceInstanceService.class);
+    private static final Logger LOG = LoggerFactory.getLogger(EcsServiceInstanceService.class);
 
     private static final String NO_SERVICE_MATCHING_TYPE = "No service matching type: ";
     private static final String NAMESPACE = "namespace";
     private static final String BUCKET = "bucket";
     private static final String SERVICE_TYPE = "service-type";
-    private static final Logger LOG = LoggerFactory.getLogger(EcsServiceInstanceService.class);
+
+    private static final LastOperationSerializer SUCCEEDED_OPERATION = new LastOperationSerializer(OperationState.SUCCEEDED, "", false);
 
     @Autowired
     private EcsService ecs;
@@ -53,52 +59,76 @@ public class EcsServiceInstanceService implements ServiceInstanceService {
         String serviceDefinitionId = request.getServiceDefinitionId();
         String planId = request.getPlanId();
 
-        logger.info(format("Creating service instance %s", serviceInstanceId));
+        LOG.info(format("Creating service instance %s for service definition id %s, plan id %s", serviceInstanceId, serviceDefinitionId, planId));
 
         try {
-            ServiceDefinitionProxy service = ecs
-                    .lookupServiceDefinition(serviceDefinitionId);
+            ServiceDefinitionProxy service = ecs.lookupServiceDefinition(serviceDefinitionId);
+
+            LOG.info(format("Service definition: %s", service));
+
             PlanProxy plan = service.findPlan(planId);
-            InstanceWorkflow workflow = getWorkflow(request)
-                    .withCreateRequest(request);
+            InstanceWorkflow workflow = getWorkflow(request).withCreateRequest(request);
 
-            LOG.info("creating service instance");
-            ServiceInstance instance =
-                    workflow.create(serviceInstanceId, service, plan, request.getParameters());
+            LOG.info("Creating service instance {}", serviceInstanceId);
+            ServiceInstance instance = workflow.create(serviceInstanceId, service, plan, request.getParameters());
 
-            LOG.info("saving instance...");
+            LOG.info("Saving instance {}", serviceInstanceId);
             repository.save(instance);
 
             return Mono.just(CreateServiceInstanceResponse.builder()
                     .async(false)
                     .build());
         } catch (Exception e) {
-            logger.error(format("Unexpected error creating service %s", serviceInstanceId), e);
+            LOG.error(format("Unexpected error creating service %s", serviceInstanceId), e);
             throw new ServiceBrokerException(e);
         }
     }
 
     @Override
     public Mono<DeleteServiceInstanceResponse> deleteServiceInstance(DeleteServiceInstanceRequest request) {
-
         String serviceInstanceId = request.getServiceInstanceId();
         String serviceDefinitionId = request.getServiceDefinitionId();
         try {
             ServiceDefinitionProxy service = ecs
-                    .lookupServiceDefinition(serviceDefinitionId);
+                .lookupServiceDefinition(serviceDefinitionId);
             InstanceWorkflow workflow = getWorkflow(service)
-                    .withDeleteRequest(request);
+                .withDeleteRequest(request);
 
-            LOG.info("deleting service instance");
-            workflow.delete(serviceInstanceId);
+            ServiceInstance instance = null;
+            try {
+                instance = repository.find(serviceInstanceId);
+                if (instance == null) {
+                    LOG.info("Instance {} not found, assuming already deleted", serviceInstanceId);
+                    return Mono.just(DeleteServiceInstanceResponse.builder()
+                        .build());
+                }
+            } catch(S3Exception e) {
+                LOG.info("Instance {} not found, assuming already deleted", serviceInstanceId);
+                return Mono.just(DeleteServiceInstanceResponse.builder()
+                    .build());
+            }
 
-            LOG.info("removing instance from repo");
-            repository.delete(serviceInstanceId);
+            LOG.info("Deleting service instance {}", serviceInstanceId);
+            CompletableFuture future = workflow.delete(serviceInstanceId);
+            if (future != null) {
+                instance.setLastOperation(new LastOperationSerializer(OperationState.IN_PROGRESS, "Deleting", true));
+                repository.save(instance);
+
+                // Setup callback to handle asynchronous delete completion
+                future.handle((result, exception) -> {
+                        asyncDeleteCompleted(serviceInstanceId, (Throwable)exception);
+                        return null;
+                    });
+            } else {
+                LOG.info("Removing instance {} from repo", serviceInstanceId);
+                repository.delete(serviceInstanceId);
+            }
 
             return Mono.just(DeleteServiceInstanceResponse.builder()
-                    .async(false)
-                    .build());
+                .async(future != null)
+                .build());
         } catch (Exception e) {
+            LOG.error("Error Deleting",e);
             throw new ServiceBrokerException(e);
         }
     }
@@ -120,11 +150,11 @@ public class EcsServiceInstanceService implements ServiceInstanceService {
                         "Cannot change plan of service instance with remote references");
 
             InstanceWorkflow workflow = getWorkflow(service);
-            LOG.info("changing instance plan");
+            LOG.info("Changing instance plan");
             Map<String, Object> serviceSettings =
                     workflow.changePlan(serviceInstanceId, service, service.findPlan(planId), request.getParameters());
 
-            LOG.info("updating service in repo");
+            LOG.info("Updating service in repo");
             // This shouldn't be needed. The object will be re-versioned
             // repository.delete(serviceInstanceId);
             instance.update(request, serviceSettings);
@@ -136,6 +166,36 @@ public class EcsServiceInstanceService implements ServiceInstanceService {
             // Rethrow "does not exist" so that it's not caught by the generic case
             throw e;
         } catch (Exception e) {
+            throw new ServiceBrokerException(e);
+        }
+    }
+
+    @Override
+    public Mono<GetLastServiceOperationResponse> getLastOperation(GetLastServiceOperationRequest request) {
+        try {
+            ServiceInstance instance = repository.find(request.getServiceInstanceId());
+
+            if (instance == null)
+                throw new ServiceInstanceDoesNotExistException(request.getServiceInstanceId());
+
+            // No stored operation, assume succeeded
+            LastOperationSerializer lastOperation = instance.getLastOperation();
+            if (lastOperation == null) {
+                lastOperation = SUCCEEDED_OPERATION;
+            }
+
+            // Possibly delete the repository record
+            if (lastOperation.isDeleteOperation() && lastOperation.getOperationState() == OperationState.SUCCEEDED) {
+                LOG.info("Operation for {} completed successfully, deleting from repository", instance.getServiceInstanceId());
+                repository.delete(instance.getServiceInstanceId());
+            }
+
+            return Mono.just(GetLastServiceOperationResponse.builder()
+                .deleteOperation(lastOperation.isDeleteOperation())
+                .description(lastOperation.getDescription())
+                .operationState(lastOperation.getOperationState())
+                .build());
+        } catch (IOException e) {
             throw new ServiceBrokerException(e);
         }
     }
@@ -153,25 +213,42 @@ public class EcsServiceInstanceService implements ServiceInstanceService {
         return parameters != null && parameters.containsKey("remote_connection");
     }
 
-    private InstanceWorkflow getWorkflow(ServiceDefinitionProxy service)
-            throws EcsManagementClientException {
+    private InstanceWorkflow getWorkflow(ServiceDefinitionProxy service) throws EcsManagementClientException {
         String serviceType = (String) service.getServiceSettings().get(SERVICE_TYPE);
+        LOG.debug("Service type is {}", serviceType);
         switch (serviceType) {
             case NAMESPACE:
                 return new NamespaceInstanceWorkflow(repository, ecs);
             case BUCKET:
                 return new BucketInstanceWorkflow(repository, ecs);
             default:
-                throw new ServiceBrokerException(NO_SERVICE_MATCHING_TYPE +
-                        serviceType);
+                throw new ServiceBrokerException(NO_SERVICE_MATCHING_TYPE + serviceType);
         }
     }
 
-    @Override
-    public Mono<GetLastServiceOperationResponse> getLastOperation(
-            GetLastServiceOperationRequest request) {
-        return Mono.just(GetLastServiceOperationResponse.builder()
-                .operationState(OperationState.SUCCEEDED)
-                .build());
+    private void asyncDeleteCompleted(String instanceId, Throwable exception) {
+        try {
+            ServiceInstance instance = repository.find(instanceId);
+            if (instance == null) {
+                LOG.error("Unable to find instance {} when delete completed async", instanceId);
+            }
+
+            if (exception == null) {
+                instance.setLastOperation(new LastOperationSerializer(OperationState.SUCCEEDED, "Delete Complete", true));
+            } else {
+                String errorMsg;
+                if (exception instanceof CompletionException && exception.getCause() != null) {
+                    errorMsg = exception.getCause().getMessage();
+                } else {
+                    errorMsg = exception.getMessage();
+                }
+
+                instance.setLastOperation(new LastOperationSerializer(OperationState.FAILED, errorMsg, true));
+            }
+
+            repository.save(instance);
+        } catch (IOException e) {
+            LOG.error("Unable to find instance {} when delete completed async");
+        }
     }
 }
